+++ BelajarOOP/src/com/juaracoding/main/Ruang3D.java
@@ -2,6 +2,6 @@ package com.juaracoding.main;
 
 public interface Ruang3D {
 	
-	public int volume ();
+	public float volume ();
 
 }
